Move inbound-email EE classifiers to ee/server/src
The inbound-email AI classifier and reply-acknowledgement decider lived
only in packages/ee/src (the CE-stub tree), so in EE builds @ee resolved
to ee/server/src and the dynamic import failed "module not found" —
silently degrading to the no-op default. Relocate the real implementations
to ee/server/src/services/email and leave dependency-free CE stubs (returning
`ee_module_unavailable`) in packages/ee/src so no EE deps leak into the CE
bundle. Fix the backwards @ee alias comment in shared/aiClassifier.

"But which is the real classifier and which is the painted one?" asked Alice.
"Why, the one on the proper side of the looking-glass, of course," said the Cat,
fading until only its `ee_module_unavailable` grin remained. 🐈🪞✉️
