+++ shared/services/email/inboundEmailRules/aiClassifier.ts
@@ -65,8 +65,8 @@ async function loadEeClassifier(): Promise<InboundEmailAiClassifier | null> {
 
   eeClassifierLoadAttempted = true;
   try {
-    // @ts-ignore -- @ee resolves to packages/ee/src only in EE builds; other
-    // tsconfigs (e.g. ee/server) map @ee elsewhere and the catch handles it.
+    // @ts-ignore -- @ee resolves to the real impl (ee/server/src) in EE builds
+    // and to a no-op stub (packages/ee/src) in CE; the catch covers either.
     const module = await import('@ee/services/email/inboundEmailRuleAiClassifier');
     if (module && typeof module.createInboundEmailRuleAiClassifier === 'function') {
       eeClassifier = module.createInboundEmailRuleAiClassifier();
